rename screens according to fsd naming
