+++ src/pages/event-page/ui/event-page/event-page.vue
