+++ src/pages/event-page/ui/event-page/index.ts
