+++ src/pages/event-page/ui/index.ts
