+++ src/pages/events-list-page/ui/events-list-page/events-list-page.vue
