+++ src/pages/events-list-page/ui/events-list-page/index.ts
