+++ src/pages/events-list-page/ui/index.ts
